Move BugCard and FishCard into Card folder and move their styles into a separate css file
