+++ public/styles/styles.css
@@ -1,18 +1,1 @@
-img {
-  width: 60px;
-}
-
-.critter-name {
-  margin: 4px 0px;
-  font-weight: 700;
-}
-
-.critter {
-  font-size: 18px;
-  background-color: #2e2e2e;
-  text-align: center;
-  width: 100%;
-  height: 100%;
-  margin: 0 4px 8px 4px;
-}
-/*# sourceMappingURL=styles.css.map */
+/* No CSS *//*# sourceMappingURL=styles.css.map */
