+++ src/Components/Cards/FishCard/index.js
@@ -1,15 +1,12 @@
 import React from "react";
+import "../index.css";
 
 class FishCard extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-
   render() {
     return (
       <div className="critter">
         <img
-          src={require(`../../data/images/Fish/${this.props.image}.png`)}
+          src={require(`../../../data/images/Fish/${this.props.image}.png`)}
           alt="fish"
         ></img>
         <p className="critter-name">{this.props.name}</p>
